REFACTOR `PdfExportValidations` to `PdfTestCase`
