+++ tests/PdfExportFromUrlTest.php
@@ -3,12 +3,9 @@
 namespace Sfneal\ViewExport\Tests;
 
 use Sfneal\ViewExport\Pdf\PdfExportService;
-use Sfneal\ViewExport\Tests\Traits\PdfExportValidations;
 
-class PdfExportFromUrlTest extends TestCase
+class PdfExportFromUrlTest extends PdfTestCase
 {
-    use PdfExportValidations;
-
     /**
      * Setup the test environment.
      *
