+++ tests/PdfExportFromViewModelTest.php
@@ -3,12 +3,10 @@
 namespace Sfneal\ViewExport\Tests;
 
 use Sfneal\ViewExport\Pdf\PdfExportService;
-use Sfneal\ViewExport\Tests\Traits\PdfExportValidations;
 use Sfneal\ViewExport\Tests\ViewModels\TestViewModel;
 
-class PdfExportFromViewModelTest extends TestCase
+class PdfExportFromViewModelTest extends PdfTestCase
 {
-    use PdfExportValidations;
 
     /**
      * Setup the test environment.
